requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
Flask==3.0.2
future==1.0.0
geocoder==1.38.1
h3==3.7.6
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pycparser==2.21
PyJWT==2.8.0
pytz==2024.1
PyWeatherKit==0.0.4
ratelim==0.1.6
requests==2.31.0
six==1.16.0
timezonefinder==6.4.1
urllib3==2.2.1
Werkzeug==3.0.1